# src/libs/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "Library",
    "description": "A simple book reservation system."
  },
  "servers": [
    {
      "url": "https://syslibrary_ts.onrender.com/",
      "description": "Production"
    },
    {
      "url": "http://localhost:3333/",
      "description": "Local"
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": ""
    },
    {
      "name": "User",
      "description": ""
    },
    {
      "name": "Book",
      "description": ""
    },
    {
      "name": "Booking",
      "description": ""
    }
  ],
  "paths": {
    "/api/v1/": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "",
        "responses": {
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Token"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            },
            "description": "Criado"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/info": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/patch": {
      "patch": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            },
            "description": "Criado"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatch"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/delete": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "content": {
              "application/json": {}
            },
            "description": "Nenhum Conteúdo"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/books/register": {
      "post": {
        "tags": [
          "Book"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            },
            "description": "Criado"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookRegister"
              }
            }
          }
        }
      }
    },
    "/api/v1/books/list/": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageQuery"
          },
          {
            "$ref": "#/components/parameters/PageSizeQuery"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/CategoryQuery"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/books/listByUser/": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/PageQuery"
          },
          {
            "$ref": "#/components/parameters/PageSizeQuery"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/CategoryQuery"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/books/update/": {
      "patch": {
        "tags": [
          "Book"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdBookQuery"
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            },
            "description": "Criado"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookPatch"
              }
            }
          }
        }
      }
    },
    "/api/v1/books/delete/": {
      "delete": {
        "tags": [
          "Book"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdBookQuery"
          }
        ],
        "responses": {
          "204": {
            "content": {
              "application/json": {}
            },
            "description": "Nenhum Conteúdo"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/books/categories/": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "",
        "parameters": [
          {
            "$ref": "#/components/parameters/SearchQuery"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookCategoryResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/bookings/register": {
      "post": {
        "tags": [
          "Booking"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            },
            "description": "Criado"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRegister"
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings/list": {
      "get": {
        "tags": [
          "Booking"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/PageQuery"
          },
          {
            "$ref": "#/components/parameters/PageSizeQuery"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "DefaultResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "message"
          }
        }
      },
      "BooleanResponse": {
        "type": "object",
        "properties": {
          "tokenIsValid": {
            "type": "string",
            "example": "boolean"
          }
        },
        "required": [
          "tokenIsValid"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "email@email.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE3MDcyMTgyOTgsImV4cCI6MTcwNzMwNDY5OCwic3ViIjoiZmQ3MmIyMzMtMzJmNi00Y2ZiLWE1ZmEtMTk0NDY3MmJiM2E0In0.JPQGJRaXsZG4UakiXbT2AZyxKmk9JhWtmFBxzZxRdTY"
          }
        },
        "required": [
          "token"
        ]
      },
      "TokenRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE3MDcyMTgyOTgsImV4cCI6MTcwNzMwNDY5OCwic3ViIjoiZmQ3MmIyMzMtMzJmNi00Y2ZiLWE1ZmEtMTk0NDY3MmJiM2E0In0.JPQGJRaXsZG4UakiXbT2AZyxKmk9JhWtmFBxzZxRdTY"
          }
        },
        "required": [
          "token"
        ]
      },
      "UserRegister": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "username"
          },
          "name": {
            "type": "string",
            "example": "name"
          },
          "email": {
            "type": "string",
            "example": "email@email.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        },
        "required": [
          "username",
          "name",
          "email",
          "password"
        ]
      },
      "UserPatch": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "username"
          },
          "name": {
            "type": "string",
            "example": "name"
          },
          "email": {
            "type": "string",
            "example": "email@email.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a0fbbdeb-fee6-4e58-bec7-8ef84e55a892"
          },
          "username": {
            "type": "string",
            "example": "username"
          },
          "name": {
            "type": "string",
            "example": "name"
          },
          "email": {
            "type": "string",
            "example": "email@email.com"
          },
          "createdAt": {
            "type": "string",
            "example": "year-mouth-day"
          },
          "updatedAt": {
            "type": "string",
            "example": "year-mouth-day"
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "email",
          "createdAt",
          "updatedAt"
        ]
      },
      "BookRegister": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "title"
          },
          "cod": {
            "type": "string",
            "example": "cod"
          },
          "editora": {
            "type": "string",
            "example": "editora"
          },
          "autor": {
            "type": "string",
            "example": "autor"
          },
          "sinopse": {
            "type": "string",
            "example": "sinopse"
          },
          "bookCategoryId": {
            "type": "string",
            "example": "729c1ed8-93bf-4e1a-9d48-d4dec11e57e7"
          },
          "qtd": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "title",
          "cod",
          "editora",
          "autor",
          "sinopse",
          "bookCategoryId",
          "qtd"
        ]
      },
      "BookPatch": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "title"
          },
          "cod": {
            "type": "string",
            "example": "cod"
          },
          "editora": {
            "type": "string",
            "example": "editora"
          },
          "autor": {
            "type": "string",
            "example": "autor"
          },
          "sinopse": {
            "type": "string",
            "example": "sinopse"
          },
          "bookCategoryId": {
            "type": "string",
            "example": "729c1ed8-93bf-4e1a-9d48-d4dec11e57e7"
          },
          "qtd": {
            "type": "number",
            "example": 1
          }
        }
      },
      "BookResponse": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "example": "a0fbbdeb-fee6-4e58-bec7-8ef84e55a892"
            },
            "title": {
              "type": "string",
              "example": "title"
            },
            "cod": {
              "type": "string",
              "example": "cod"
            },
            "editora": {
              "type": "string",
              "example": "editora"
            },
            "autor": {
              "type": "string",
              "example": "autor"
            },
            "sinopse": {
              "type": "string",
              "example": "sinopse"
            },
            "bookCategoryId": {
              "type": "string",
              "example": "729c1ed8-93bf-4e1a-9d48-d4dec11e57e7"
            },
            "qtd": {
              "type": "number",
              "example": 1
            },
            "idUser": {
              "type": "string",
              "example": "a0fbbdeb-fee6-4e58-bec7-8ef84e55a892"
            },
            "createdAt": {
              "type": "string",
              "example": "year-mouth-day"
            },
            "updatedAt": {
              "type": "string",
              "example": "year-mouth-day"
            }
          },
          "required": [
            "id",
            "title",
            "cod",
            "editora",
            "autor",
            "sinopse",
            "bookCategoryId",
            "qtd",
            "idUser",
            "createdAt",
            "updatedAt"
          ]
        }
      },
      "BookCategoryResponse": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "example": "a0fbbdeb-fee6-4e58-bec7-8ef84e55a892"
            },
            "name": {
              "type": "string",
              "example": "name"
            },
            "createdAt": {
              "type": "string",
              "example": "year-mouth-day"
            },
            "updatedAt": {
              "type": "string",
              "example": "year-mouth-day"
            }
          },
          "required": [
            "id",
            "name",
            "createdAt",
            "updatedAt"
          ]
        }
      },
      "BookingRegister": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string",
            "example": "a0fbbdeb-fee6-4e58-bec7-8ef84e55a892"
          }
        },
        "required": [
          "bookId"
        ]
      },
      "BookingResponse": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "example": "a0fbbdeb-fee6-4e58-bec7-8ef84e55a892"
            },
            "idUser": {
              "type": "string",
              "example": "a0fbbdeb-fee6-4e58-bec7-8ef84e55a892"
            },
            "bookId": {
              "type": "string",
              "example": "a0fbbdeb-fee6-4e58-bec7-8ef84e55a892"
            },
            "createdAt": {
              "type": "string",
              "example": "year-mouth-day"
            },
            "updatedAt": {
              "type": "string",
              "example": "year-mouth-day"
            }
          },
          "required": [
            "id",
            "idUser",
            "bookId",
            "createdAt",
            "updatedAt"
          ]
        }
      }
    },
    "parameters": {
      "IdBookQuery": {
        "name": "id",
        "in": "query",
        "schema": {
          "type": "string"
        }
      },
      "Token": {
        "name": "token",
        "in": "query",
        "schema": {
          "type": "string"
        }
      },
      "PageQuery": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer"
        }
      },
      "PageSizeQuery": {
        "name": "pageSize",
        "in": "query",
        "schema": {
          "type": "integer"
        }
      },
      "SearchQuery": {
        "name": "search",
        "in": "query",
        "schema": {
          "type": "string"
        }
      },
      "CategoryQuery": {
        "name": "categoryId",
        "in": "query",
        "schema": {
          "type": "string"
        }
      }
    }
  }
}